refactor: handle dependencies with nested directory
