+++ pkg/devspace/compose/testdata/build_network/devspace.yaml
@@ -4,7 +4,6 @@ name: docker-compose
 images:
   foo:
     image: foo
-    dockerfile: Dockerfile
     network: host
 
 deployments:
