+++ pkg/devspace/compose/testdata/build_target/devspace.yaml
@@ -4,7 +4,6 @@ name: docker-compose
 images:
   foo:
     image: foo
-    dockerfile: Dockerfile
     target: dev
 
 deployments:
