+++ pkg/devspace/compose/testdata/depends_on_with_build/backend/Dockerfile
@@ -1,0 +1,1 @@
+FROM alpine
